backend/requirements.txt
aiohttp==3.9.1
aiosqlite==0.19.0
asyncpg==0.29.0
numpy==1.26.2
pandas==2.2.0
psutil==5.9.7
redis==5.0.1
scikit-learn==1.3.2
tenacity==8.2.3
pytest==7.4.3
pytest-aiohttp==1.0.4
pytest-asyncio==0.23.2
joblib==1.3.2
uvicorn==0.24.0
fastapi==0.104.1
python-dotenv==1.0.1
prometheus_client==0.19.0
flask==3.0.2
flask-cors==4.0.0
flask-socketio==5.3.6
pyyaml==6.0.2
sqlalchemy==2.0.23
pydantic==2.5.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6